feat: Begin to introduce save/load functionality
This fulfils the requirements of writing and reading
from a file. The save seems to work fine, however the
loading still has issues. Something that I changed
to fix an error in loading has broken the inventory
in the Character sheet.
I suspect that this may be a result of passing around
inventory arrays of pointers, and then destructors
remove what I don't expect... Or something.